Move selection components into a folder
